requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.3
greenlet==3.1.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
playwright==1.48.0
pluggy==1.6.0
pyee==12.0.0
pytest==8.3.3
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.5.2
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.5
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.5.0
